requirements.txt
# Core ML Libraries
torch==2.1.0
transformers==4.36.0

# Data Processing
numpy==1.24.0
pandas==2.1.0
scipy==1.11.0
scikit-learn==1.3.0

# LLM and RAG Framework
openai>=1.10.0
langchain
langchain-openai
langchain-experimental
langchain-community

# Vector Stores
chromadb==0.4.18
faiss-cpu==1.7.4
docarray==0.39.1

# Evaluation
rouge-score==0.1.2
datasets==2.14.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.0

# Visualization
matplotlib==3.8.0
jupyter==1.0.0

# Development Tools
pytest==7.4.0
black==23.12.0
flake8==6.1.0
mypy==1.7.0
